+++ crates/cgp-core/src/prelude.rs
@@ -1,4 +1,4 @@
-pub use cgp_async::{async_trait, Async};
+pub use cgp_async::{async_trait, Async, MaybeSend, MaybeStatic, MaybeSync};
 pub use cgp_component::{
     define_components, delegate_components, derive_component, DelegateComponent, HasComponents,
 };