+++ crates/cgp-inner/src/lib.rs
@@ -2,12 +2,11 @@
 
 extern crate alloc;
 
-use cgp_async::Async;
 use cgp_component::{derive_component, DelegateComponent, HasComponents};
 
 #[derive_component(InnerComponent, ProvideInner<Context>)]
-pub trait HasInner: Async {
-    type Inner: Async;
+pub trait HasInner {
+    type Inner;
 
     fn inner(&self) -> &Self::Inner;
 }